Prevent type checks on adminApp from external apps
