+++ apps/image-generation/src/client.ts
@@ -4,9 +4,9 @@ import { hc } from 'hono/client'
 export {
   TrpFrogImageGenerationResultSchema,
   type TrpFrogImageGenerationResult,
-} from './trpfrog-diffusion/generateNew'
+} from './trpfrog-diffusion/schema'
 
-import type { AppType } from '.'
+import type { AppType } from './app'
 
 export function createTrpFrogImageGenerationClient(env: 'development' | 'production' | 'test') {
   return hc<AppType>(endpoints(env).imageGeneration)
